fix: move migration modules to src instead of skills
