+++ .agents/skills/cxas-dfcx-migration/scripts/_bundle.py
@@ -6,145 +6,21 @@
 #
 #     https://www.apache.org/licenses/LICENSE-2.0
 
-"""IR bundle persistence — single source of truth for cross-script state.
+"""Skill-local re-export of the promoted IR bundle module.
 
-The IR bundle is the artifact that lets `migrate.py`, `stage1.py`, and
-`stage2.py` run as independent processes against the same migration. It
-contains everything the downstream stages need to skip the expensive
-per-flow Step 2A/2B/2C compile pipeline:
+The implementation lives in :mod:`cxas_scrapi.migration.ir_bundle` so the
+same logic is reachable from the CLI (``cxas migrate dfcx-cxas stage*``)
+and any other caller. Skill scripts continue to ``import _bundle`` so
+callsites don't need to change."""
 
-* The original `MigrationConfig` (project, location, model, etc.)
-* The source `DFCXAgentIR` (needed for tool-mock context in Stage 2)
-* The current `MigrationIR` (target — mutated by each stage)
-* Stage history + version checkpoints for the audit report.
-"""
-
-from __future__ import annotations
-
-import glob
-import logging
-import os
-from datetime import datetime
-from typing import Any
-
-from pydantic import BaseModel, Field
-
-from cxas_scrapi.migration.data_models import (
-    DFCXAgentIR,
-    MigrationConfig,
-    MigrationIR,
+from cxas_scrapi.migration.ir_bundle import (  # noqa: F401
+    IRBundle,
+    StageHistoryEntry,
+    append_stage,
+    attach_grouping,
+    attach_version,
+    find_default_bundle,
+    load,
+    save,
+    save_for_target,
 )
-
-logger = logging.getLogger(__name__)
-
-
-class StageHistoryEntry(BaseModel):
-    phase: str  # "migrate", "stage1", "stage2"
-    status: str  # "ok", "fail", "partial"
-    started_at: datetime
-    ended_at: datetime | None = None
-    notes: str = ""
-
-
-class IRBundle(BaseModel):
-    """Persisted state shared across migrate / stage1 / stage2."""
-
-    schema_version: str = "2"
-    created_at: datetime = Field(default_factory=datetime.now)
-    config: MigrationConfig
-    source_agent_data: DFCXAgentIR
-    ir: MigrationIR
-    stage_history: list[StageHistoryEntry] = Field(default_factory=list)
-    app_url: str | None = None
-    version_checkpoints: list[tuple[str, str]] = Field(default_factory=list)
-    grouping: dict[str, Any] | None = (
-        None  # populated when Stage 1 consolidates
-    )
-    # Snapshot of `ir` BEFORE consolidation mutated agents/tools, kept on disk
-    # so the original 1:1 IR survives even after the consolidation flow
-    # collapses N agents into M groups.
-    pre_consolidation_ir: MigrationIR | None = None
-
-
-# ---------------------------------------------------------------------------
-# Disk I/O
-# ---------------------------------------------------------------------------
-
-
-def _bundle_filename(target_name: str) -> str:
-    return f"{target_name}_ir.json"
-
-
-def save(bundle: IRBundle, path: str) -> str:
-    """Atomic write — write to a tempfile then rename."""
-    tmp = path + ".tmp"
-    with open(tmp, "w") as f:
-        f.write(bundle.model_dump_json(indent=2))
-    os.replace(tmp, path)
-    logger.info("IR bundle saved → %s", path)
-    return path
-
-
-def save_for_target(
-    bundle: IRBundle, target_name: str, cwd: str | None = None
-) -> str:
-    path = os.path.join(cwd or os.getcwd(), _bundle_filename(target_name))
-    return save(bundle, path)
-
-
-def load(path: str) -> IRBundle:
-    with open(path) as f:
-        return IRBundle.model_validate_json(f.read())
-
-
-def find_default_bundle(
-    target_name: str | None = None, cwd: str | None = None
-) -> str | None:
-    """Locate an IR bundle on disk.
-
-    If `target_name` is supplied: returns `<cwd>/<target_name>_ir.json` if it
-    exists, else None.
-    Otherwise: returns the newest `*_ir.json` in cwd, or None.
-    """
-    cwd = cwd or os.getcwd()
-    if target_name:
-        candidate = os.path.join(cwd, _bundle_filename(target_name))
-        return candidate if os.path.exists(candidate) else None
-    matches = glob.glob(os.path.join(cwd, "*_ir.json"))
-    if not matches:
-        return None
-    matches.sort(key=os.path.getmtime, reverse=True)
-    return matches[0]
-
-
-# ---------------------------------------------------------------------------
-# Convenience mutators
-# ---------------------------------------------------------------------------
-
-
-def append_stage(
-    bundle: IRBundle,
-    phase: str,
-    status: str,
-    started_at: datetime,
-    notes: str = "",
-) -> None:
-    bundle.stage_history.append(
-        StageHistoryEntry(
-            phase=phase,
-            status=status,
-            started_at=started_at,
-            ended_at=datetime.now(),
-            notes=notes,
-        )
-    )
-
-
-def attach_version(
-    bundle: IRBundle, display_name: str, description: str
-) -> None:
-    bundle.version_checkpoints.append((display_name, description))
-
-
-def attach_grouping(bundle: IRBundle, groupings: dict[str, Any]) -> None:
-    bundle.grouping = groupings
